+++ R/epidatatools-package.R
@@ -4,5 +4,6 @@
 utils::globalVariables(c("bin_value", "bin_number", "cdf", "name", "value"))
 
 ## usethis namespace: start
+#' @importFrom lifecycle deprecated
 ## usethis namespace: end
 NULL